Keep XP fractional in the backend; show whole numbers (nearest) in the UI
Progression rounded XP twice per grant (after the Facilities multiplier,
then again after the Coaches bonus), so precision was lost on every grant.
XP is now stored and calculated as a float end to end, and only the UI
turns it into an integer.

Backend
- PlayerProgressionState.current_xp and XpHistoryEntry.xp_gained are float.
- ApplyXpBonuses no longer rounds at either step (the minimum-1 floor stays).
- Persisted JSON and the xp_updated event payload carry the fractional
  value, cleaned of float noise (15.45, not 15.449999809).
- Tier lookup now picks the highest tier whose minimum has been reached.
  The old inclusive whole-number ranges (0-49, 50-99, ...) left fractional
  XP such as 99.5 matching no tier, and any XP above the last range's max
  fell back to rookie; both are fixed by this.

Display
- New XpFormat.ToDisplay is the single place that converts to an integer:
  the nearest whole number, with .5 rounding up (plain Math.Round rounds
  half to even, which would show 30.5 as 30). The running total is rounded
  once, so two 15.4 gains show 31 (30.8), not 15 + 15.
- It strips float noise first: 100 x 1.08f is 108.0000076 in float math,
  and a value that is 30.5 up to noise must land on the .5 side.
- SeasonManager.PlayerXP / LastXpGained return the display integer, which
  covers the hub, bracket, rewards title and sim-feedback screens without
  touching each one. PlayerXPExact is added for logic; RewardsUI's tier
  unlock now uses it so it agrees with ProgressionService's own tiering.
- XPUI (chart/list/header), the CCAS acquisition hub, the legacy
  PlayerProgressionUI and the Facilities ProgressionSummaryPanel are
  updated. The Facilities panel read the JSON with Value<int?>, which would
  have applied a different rounding rule from the rest of the UI.

Tests (203 assertions across 7 suites, all passing)
- New: display rounding using the spec's worked examples plus half-up and
  float-noise cases, fractional accumulation without drift (ten 1-XP grants
  total 10 x multiplier, not 10), persistence round trip, and tier lookup
  on fractional and above-range XP.
- Updated the assertions that had baked in the old rounding (a 7-XP
  match_win is now 7.21, the season-reward and Facilities expectations).
  CCASUnitTests.GetStateXp returned int and needed a one-line conversion.

Known edge (not changed): tier is decided on the exact value while the UI
shows the rounded value, so XP in [49.5, 50) displays as 50 but is still
rookie until it actually reaches 50.

+++ Assets/Scripts/Progression/Models/PlayerModels.cs
@@ -23,13 +23,13 @@ public class XpHistoryEntry
     public string id;
     public string player_id;
     public string timestamp;
-    public int xp_gained;
+    public float xp_gained;
     public string source;
 
     public XpHistoryEntry() {}
 
-    
-    public XpHistoryEntry(string playerId, int xpGained, string source)
+
+    public XpHistoryEntry(string playerId, float xpGained, string source)
     {
         id = Guid.NewGuid().ToString();
         player_id = playerId;
@@ -45,7 +45,7 @@ public class XpHistoryEntry
 public class PlayerProgressionState
 {
     public string player_id;
-    public int current_xp;
+    public float current_xp;
     public string current_tier;
     public List<XpHistoryEntry> xp_history;
